fix: GPS filter — float→double, spike bypass, mutex, duplicate detection
4 bugs corrigés dans MapGPSFilter :
1. float→double pour lat/lon (catastrophic cancellation à 7 digits)
2. Spike rejection après gap >120s (téléportation silencieuse)
3. Mutex mutable pour accès cross-core (défense en profondeur)
4. Détection doublon lat/lon (évite retraitement même trame NMEA)

Delta diagnostique déplacé AVANT la mise à jour du filtre pour
refléter l'écart réel GPS brut vs position filtrée.

+++ src/map_input.cpp
@@ -126,7 +126,7 @@ namespace MapInput {
     void btn_map_recenter_clicked(lv_event_t* e) {
         ESP_LOGI(TAG, "Recentering on GPS");
         map_follow_gps = true;
-        float initLat, initLon;
+        double initLat, initLon;
         if (gpsFilter.getUiPosition(&initLat, &initLon)) {
             ESP_LOGI(TAG, "Recentered on GPS: %.4f, %.4f", initLat, initLon);
         } else {
